+++ OthelloTypes.h
@@ -25,7 +25,7 @@
 #include <unordered_set>
 
 /* Macros and Defines */
-#define VERSION "0.33.8"
+#define VERSION "0.33.9"
 
 /* Compression mode for RSF output files. */
 #define COMPRESS_NONE       0   /* all files uncompressed (.rsf)                              */